pyproject.toml
[project]
name = "wayland-secret-insert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygobject>=3.56.0",
    "python-wofi>=0.3.1",
    "typer>=0.24.1",
]
